ObjStorage: Return GitObjectResult instead of GitObject?
Fetching an object can fail for so so many reasons. Just returning null
doesn't give the caller the reason why it failed. Ideally, it would have
been nice to have golang style error handling, because that is what I'm
creating in the end.

(GoLang/Result/Swift)

This is the first commit in the long number of commits to refactor the
entire API.