ObjectStorage: Catch all exceptions
I don't want exceptions ever bubbling out of the ObjectStorage. Ideally,
when using DartGit you'll get `Result<Type>` return types, which will
have the exception if it is fired.

This way, I'm forced to handle the exceptions in GitJournal, and it's
easy to handle them as I don't need to put a try and catch block around
every statement.