ObjectStorage: Avoid using read
Instead use readType instead. This way we avoid the extra cast, and
there is less chance of making an invalid cast and getting a runtime
exception.

Also, this way the error messages can be much better if someone does try
to read a commit when the object is a blob, for example.

This caught one bug in diff-tree.