Speedup by relying on filesystem ctime/mtime/filesize
On my test repo with a few text files and some images, but not too large
(194 MB in total) this speeds "add" and "commit" up as follows:

* When the command is "add", this takes 2 seconds without my ctime/mtime patch and ~0.2 without.
* When the command is "commit", this takes 2 seconds but only due to asserts that get removed for release builds..

The use-case I have in mind is that GitJournal takes 30 seconds to
startup on my phone, using this same folder of notes files. This commit
makes that usable again.