Improve merge algorithm
Handle the case where we have -

base - 1.md 2.md
a - 1.md 2.md 3.md
b - 1.md

In this case the file was deleted in 2.md and should not appear in the
final merged entries. This represents a common case where the file was
deleted in the remote branch but not locally.

Fixes https://github.com/GitJournal/GitJournal/issues/962